requirements.txt
stravalib
python-dotenv
pandas
matplotlib
fpdf